Fixed signed/unsigned comparison warning on older linux machines.
git-svn-id: http://svn.gplates.org/projects/gplates/branches/platepolygon-2008-08-15@6032 e74ed31e-5968-4cd8-ac0b-be5e0f3b170c

+++ src/gui/TopologyTools.cc
@@ -2838,7 +2838,8 @@ return false;
 
 	// Get index of next section.
 	int next_section_index = curr_section_index + 1;
-	if (next_section_index == d_section_ranges_into_topology_vertices.size())
+	if (next_section_index ==
+			boost::numeric_cast<int>(d_section_ranges_into_topology_vertices.size()))
 	{
 		next_section_index = 0;
 	}
